Creates common space between media and clinicalsummary
