--- src/main/java/redox/datamodel/clinicalsummary/common/Address.java
@@ -1,109 +1,0 @@
-
-package redox.datamodel.clinicalsummary.common;
-
-import java.util.HashMap;
-import java.util.Map;
-import com.fasterxml.jackson.annotation.JsonAnyGetter;
-import com.fasterxml.jackson.annotation.JsonAnySetter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-    "StreetAddress",
-    "City",
-    "State",
-    "ZIP",
-    "County",
-    "Country"
-})
-public class Address {
-
-    @JsonProperty("StreetAddress")
-    private String streetAddress;
-    @JsonProperty("City")
-    private String city;
-    @JsonProperty("State")
-    private String state;
-    @JsonProperty("ZIP")
-    private String zIP;
-    @JsonProperty("County")
-    private String county;
-    @JsonProperty("Country")
-    private String country;
-    @JsonIgnore
-    private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    @JsonProperty("StreetAddress")
-    public String getStreetAddress() {
-        return streetAddress;
-    }
-
-    @JsonProperty("StreetAddress")
-    public void setStreetAddress(String streetAddress) {
-        this.streetAddress = streetAddress;
-    }
-
-    @JsonProperty("City")
-    public String getCity() {
-        return city;
-    }
-
-    @JsonProperty("City")
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    @JsonProperty("State")
-    public String getState() {
-        return state;
-    }
-
-    @JsonProperty("State")
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    @JsonProperty("ZIP")
-    public String getZIP() {
-        return zIP;
-    }
-
-    @JsonProperty("ZIP")
-    public void setZIP(String zIP) {
-        this.zIP = zIP;
-    }
-
-    @JsonProperty("County")
-    public String getCounty() {
-        return county;
-    }
-
-    @JsonProperty("County")
-    public void setCounty(String county) {
-        this.county = county;
-    }
-
-    @JsonProperty("Country")
-    public String getCountry() {
-        return country;
-    }
-
-    @JsonProperty("Country")
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    @JsonAnyGetter
-    public Map<String, Object> getAdditionalProperties() {
-        return this.additionalProperties;
-    }
-
-    @JsonAnySetter
-    public void setAdditionalProperty(String name, Object value) {
-        this.additionalProperties.put(name, value);
-    }
-
-}
