+++ src/main/java/redox/datamodel/clinicalsummary/patientquery/PatientQuery.java
@@ -9,9 +9,9 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import redox.datamodel.clinicalsummary.common.Location;
-import redox.datamodel.clinicalsummary.common.Meta;
-import redox.datamodel.clinicalsummary.common.Patient;
+import redox.datamodel.common.Location;
+import redox.datamodel.common.Meta;
+import redox.datamodel.common.Patient;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
