+++ src/main/java/redox/datamodel/clinicalsummary/patientqueryresponse/Encounter.java
@@ -12,6 +12,9 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import redox.datamodel.clinicalsummary.common.*;
+import redox.datamodel.common.Identifier;
+import redox.datamodel.common.Location;
+import redox.datamodel.common.Provider;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
