+++ src/main/java/redox/datamodel/clinicalsummary/patientqueryresponse/PatientQueryResponse.java
@@ -11,6 +11,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import redox.datamodel.clinicalsummary.common.*;
+import redox.datamodel.common.Meta;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
