+++ src/main/java/redox/datamodel/common/Patient.java
@@ -1,5 +1,5 @@
 
-package redox.datamodel.media.common;
+package redox.datamodel.common;
 
 import java.util.HashMap;
 import java.util.List;
@@ -10,19 +10,20 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import redox.datamodel.media.common.Demographics;
-import redox.datamodel.media.common.Identifier;
+import redox.datamodel.common.Demographics;
+import redox.datamodel.common.Identifier;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-    "Identifiers",
-    "Demographics",
-    "Notes"
+        "Identifiers",
+        "Demographics",
+        "Notes"
 })
 public class Patient {
 
     @JsonProperty("Identifiers")
     private List<Identifier> identifiers = null;
+    //optional
     @JsonProperty("Demographics")
     private Demographics demographics;
     @JsonProperty("Notes")
