+++ src/main/java/redox/datamodel/common/Source.java
@@ -1,5 +1,5 @@
 
-package redox.datamodel.clinicalsummary.common;
+package redox.datamodel.common;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -15,7 +15,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
     "ID",
     "Name"
 })
-public class Destination {
+public class Source {
 
     @JsonProperty("ID")
     private String iD;
