+++ src/main/java/redox/datamodel/media/newmedia/NewMedia.java
@@ -10,8 +10,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import redox.datamodel.media.common.Media;
-import redox.datamodel.media.common.Meta;
-import redox.datamodel.media.common.Patient;
+import redox.datamodel.common.Meta;
+import redox.datamodel.common.Patient;
 import redox.datamodel.media.common.Visit;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
