+++ src/main/java/redox/datamodel/media/replace/Replace.java
@@ -9,8 +9,8 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import redox.datamodel.media.common.Meta;
-import redox.datamodel.media.common.Patient;
+import redox.datamodel.common.Meta;
+import redox.datamodel.common.Patient;
 import redox.datamodel.media.common.Visit;
 import redox.datamodel.media.common.Media;
